+++ changelog.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v2.3.0
+
+- Added uniqueness argument to gene_cluster()
+
+## v2.2.0
+
+- Fix #182. MCL compilation issue.
+- Refactored gene_cluster() function. Added as_list argument. The default value for cluster is now NULL (0 previously).
+
 ## v2.1.0
 
 - Refactored usage of do.call() in gm_report().